chore: remove obsoleted acapy interop fixes
