--- pkg/didcomm/protocol/decorator/decorator_interop.go
@@ -1,13 +1,0 @@
-// +build ACAPyInterop
-
-/*
-Copyright SecureKey Technologies Inc. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package decorator
-
-const (
-	doACAPyInterop = true
-)
